Fix for #200
