+++ deploy_fake_OCEAN.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python
+#
+# Copyright 2021 Ocean Protocol Foundation
+# SPDX-License-Identifier: Apache-2.0
+#
+
+from ocean_lib.ocean.deploy import deploy_fake_OCEAN
+
+if __name__ == "__main__":
+    deploy_fake_OCEAN()
